Initial checkin for experimental work on the Windows port of USD.
This branch should be considered a work-in-progress and may not
build or work as expected.

+++ pxr/base/lib/js/json.cpp
@@ -233,7 +233,7 @@ JsParseString(
                     nlpos = i;
                 }
             }
-            error->column = eoff - nlpos;
+            error->column = static_cast<unsigned int>(eoff - nlpos);
             error->reason = rj::GetParseError_En(result.Code());
         }
         return JsValue();
