+++ pxr/usd/lib/pcp/composeSite.cpp
@@ -73,7 +73,7 @@ PcpComposeSiteReferences( const PcpLayerStackSite & site,
     SdfReferenceListOp curListOp;
 
     result->clear();
-    for (int i = layers.size() - 1; i >= 0; --i) {
+    for (int i = static_cast<int>(layers.size()) - 1; i >= 0; --i) {
         const SdfLayerHandle& layer = layers[i];
         if (layer->HasField(site.path, field, &curListOp)) {
             const SdfLayerOffset* layerOffset =
